Xart added 3 options for fansite posters.
if it makes it to option 3 the posters pulled are blank or empty. Something is wrong with the way we are pulling posters from xartbeauties despite it working on every other fansite i've tried.

The code works so im leaving it in and will create an issue to discuss.

As i work through my xart collection im adding manual fixes to some titles which googlesearch doesn't work for.